refactor: 온보딩 설명 페이지 컴포넌트 분리 및 폴더 구조 변경
